# apps/inngest-app/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inngest App API",
    "version": "1.0.0",
    "description": "API documentation for the Inngest app"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your bearer token in the format: Bearer {token}"
      }
    },
    "schemas": {
...
      "CredentialUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "secret": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "accessToken": {
                "type": "string"
              },
              "refreshToken": {
                "type": "string"
              },
              "expiresIn": {
                "type": "number"
              },
              "scopes": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          },
          "config": {
            "type": "object"
          }
        }
      },
...
    },
...
  },
  "paths": {
    "/api/credentials/{id}": {
      "get": {
        "summary": "Get a credential",
        "description": "Returns a single credential for the authenticated user.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A single credential."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Credential not found."
          },
          "500": {
            "description": "Error getting credential."
          }
        }
      },
      "put": {
        "summary": "Update a credential that is of type ApiKey",
        "tags": [
          "Credentials"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Updates a single credential for the authenticated user.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CredentialUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated credential."
          },
          "400": {
            "description": "Missing credential in request body."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Error updating credentials."
          }
        }
      },
      "delete": {
        "summary": "Delete a credential",
        "description": "Deletes a single credential for the authenticated user.",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Credentials deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Error deleting credentials."
          }
        }
      }
    },
    "/api/credentials/google/auth-url": {
      "get": {
        "summary": "Get Google OAuth URL from specified scopes",
        "tags": [
          "Google"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "scopes",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "required": true
          }
        ],
        "description": "Returns a Google OAuth URL for the authenticated user.",
        "responses": {
          "200": {
            "description": "A Google OAuth URL."
          },
          "401": {
            "description": "Unauthorized."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
    "/api/credentials/google/callback": {
      "get": {
        "summary": "Google OAuth callback",
        "description": "Handles the Google OAuth callback.",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirects to the integrations page."
          },
          "400": {
            "description": "Missing code or state from Google OAuth callback."
          },
          "500": {
            "description": "Internal Server Error."
          }
        }
      }
    },
...
  },
...
}